# lib/api-zod/src/generated/types/bonusClaimResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BonusClaimResponse {
  /** Whether the bonus entitlement is recorded for this device */
  granted: boolean;
  /** Whether the bonus was newly granted by this call (true) or already claimed previously (false) */
  newlyGranted: boolean;
}