# lib/api-zod/src/generated/types/checkBonusRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckBonusRequest {
  /**
     * Stable device identifier used to track bonus entitlement
     * @maxLength 128
     */
  deviceId: string;
}