# lib/api-zod/src/generated/types/statsAndHistoryResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatsResponse } from './statsResponse';
import type { UsageEventResponse } from './usageEventResponse';

export interface StatsAndHistoryResponse {
  stats: StatsResponse;
  /** Chronological event history, newest first */
  history: UsageEventResponse[];
}