# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UsageStatsBucket } from './usageStatsBucket';

export interface StatsResponse {
  today: UsageStatsBucket;
  thisWeek: UsageStatsBucket;
  allTime: UsageStatsBucket;
  /** Total number of recorded events */
  totalEvents: number;
}