# lib/api-zod/src/generated/types/tarotCard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TarotCard {
  /** Name of the tarot card */
  name: string;
  /** Core theme of the card */
  theme: string;
  /** Upright interpretation */
  upright: string;
  /** Reversed interpretation */
  reversed: string;
  /** Position in the spread (e.g. Passé, Présent, Avenir) */
  position: string;
  /** Whether the card is reversed */
  isReversed: boolean;
}