# lib/api-zod/src/generated/types/tarotDetailReadingResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TarotDetailReadingResponse {
  /** Detailed ~200 word interpretation explaining each card */
  longReading: string;
  /** Brief guidance from the reading */
  advice: string;
  /** Short "what to do now" action */
  nextSteps: string;
  /**
     * Credits consumed by this request
     * @minimum 0
     */
  creditsUsed: number;
  /**
     * Time spent generating the AI response in milliseconds
     * @minimum 0
     */
  generationTimeMs: number;
}