# lib/api-zod/src/generated/types/tarotFollowUpDetailResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TarotFollowUpDetailResponse {
  /** Detailed ~200 word interpretation answering the follow-up question */
  longReading: string;
  /** Brief guidance from the follow-up reading */
  advice: string;
  /** Short "what to do now" action */
  nextSteps: string;
  /**
     * Credits consumed by this request
     * @minimum 0
     */
  creditsUsed: number;
  /**
     * Time spent generating the AI response in milliseconds
     * @minimum 0
     */
  generationTimeMs: number;
}