# lib/api-zod/src/generated/types/tarotFollowUpQuestionsRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TarotCard } from './tarotCard';
import type { TarotReadingContext } from './tarotReadingContext';

export interface TarotFollowUpQuestionsRequest {
  /** The user's original question */
  question: string;
  /**
     * @minItems 3
     * @maxItems 3
     */
  cards: TarotCard[];
  context: TarotReadingContext;
}