# lib/api-zod/src/generated/types/tarotFollowUpQuestionsResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TarotFollowUpQuestionsResponse {
  /**
     * Three follow-up questions suggested by the AI
     * @minItems 3
     * @maxItems 3
     */
  questions: string[];
  /**
     * Credits consumed by this request
     * @minimum 0
     */
  creditsUsed: number;
  /**
     * Time spent generating the AI response in milliseconds
     * @minimum 0
     */
  generationTimeMs: number;
}