# lib/api-zod/src/generated/types/tarotReadingContext.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TarotReadingContext {
  /** The previously generated short reading */
  shortReading: string;
  /** The previously generated detailed reading */
  longReading: string;
  /** The previously generated advice */
  advice: string;
  /** The previously generated next steps */
  nextSteps: string;
}